backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
python-multipart==0.0.6
cryptography==40.0.2
pyOpenSSL==23.1.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.30.0
aiofiles==23.1.0
httpx==0.24.0
